Time step counter instead of time stamp integral files.

+++ core/src/io/io.h
@@ -8,4 +8,4 @@ int calc_delta_t(double, double *, Grid *);
 int interpolation_t(State *, State *, State *, double, double, double);
 int set_init_data(char[], State *);
 int write_out(State *, double [], int, double, double, Diagnostics *, Forcings *, Grid *, Dualgrid *, char [], Io_config *);
-int write_out_integral(State *, double, char [], Grid *, Dualgrid *, int);
+int write_out_integral(State *, int, char [], Grid *, Dualgrid *, int);
